Rework as a test in function_generator instead
